Add the ability to run a playbook
The playbook is currently in JSON format. A playbook is an object
with `vars` and `tasks`. `tasks` is a list of modules. To run
a playbook means that the JSON is parsed, and modules defined by
the tasks are loaded, then each task is run by the module. If any
task fails, the the program exits with a status of 1.